sim: features use retail's types too -- the struct I had not audited
Asked whether the data-type work was finished. It was not: Feature and FeatType
had never been looked at, and two of their fields are folded into the checksum
as raw float bits.

  * Feature x/z -> Fixed. Every other position in the sim is 16.16; these were
    the last floats holding world coordinates, and they decide reclaim geometry
    and nav blocking.
  * Feature work/workFull -> Fixed, and the reclaim drain is now an exact
    per-tick constant (kReclaimRate is work/second) rather than a dt multiply,
    matching what the corpse drain already does.
  * Feature dmg -> int32, and FeatType hp -> int32. Retail reads a feature's
    `damage` with readInt, and weapon damage is an int as well, so both sides of
    "has this feature absorbed enough to die" are integers now.
  * Order clickX/clickZ and AbandonedGoal x/z -> Fixed; both are positions.

FeatType::energy stays a float BECAUSE RETAIL'S IS: the feature TDF's `energy`
resolves to readFloat, where `damage`, `reclaimable` and `blocking` are readInt.
Per field, as always.

I had this one backwards earlier. A pattern pass once "corrected" `ft.hp <= 0`
to Fixed() and I reverted it with the note "FeatType::hp is a float (TDF
damage=)" -- which described OUR type and not retail's. The binary says int.

THE CHECKSUM MOVES, and here that is the point rather than a side effect:
e85c65c9249c914b -> 8f42b8ce04e60a86, stable over two runs. The feature fold
was memcpy'ing the BITS of a float work value; it now folds the raw 16.16 and
the integer dmg. Unlike every other conversion in this run, this one changes
what is hashed rather than just how it is stored.

All 14 tests pass.

+++ src/client/gameview_hud.cpp
@@ -120,7 +120,7 @@
                     if (useSimThread_) lk.lock();
                     for (const auto& f : world_.features()) {
                         if (!f.alive) continue;
-                        float dx = f.x - wx, dz = f.z - wz, d = dx * dx + dz * dz;
+                        float dx = f.x.toFloat() - wx, dz = f.z.toFloat() - wz, d = dx * dx + dz * dz;
                         float r = 18.0f + 8.0f * float(std::max(f.fx, f.fz));
                         if (d < r * r && d < bestF) { bestF = d; fid = f.id; fhit = true; }
                     }
@@ -649,7 +649,7 @@
                 if (useSimThread_) lk.lock();
                 for (const auto& f : world_.features()) {
                     if (!f.alive) continue;
-                    float dx = f.x - wx, dz = f.z - wz, d = dx * dx + dz * dz;
+                    float dx = f.x.toFloat() - wx, dz = f.z.toFloat() - wz, d = dx * dx + dz * dz;
                     float r = 18.0f + 8.0f * float(std::max(f.fx, f.fz));
                     if (d < r * r && d < bestF) { bestF = d; fid = f.id; fhit = true; }
                 }
