+++ src/client/gameview_input.cpp
@@ -393,7 +393,7 @@
                 std::unique_lock<std::mutex> lk(simMutex_, std::defer_lock);
                 if (useSimThread_) lk.lock();
                 for (const auto& f : world_.features())
-                    if (f.id == fid) { fx = f.x; fz = f.z; break; }
+                    if (f.id == fid) { fx = f.x.toFloat(); fz = f.z.toFloat(); break; }
             }
             float dx = fx - (b ? b->x : x), dz = fz - (b ? b->z : z);
             order.push_back({dx * dx + dz * dz, fid});
